advance validation rules, improve Range, improve SystemState
Range has a new pattern validation strategy. Though this code is rem'd out right now because the validation is happening on the form so it doesn't need to be done twice.

System state got it's first new 'group' checker, so instead of saying "is this value in the array (get array from SystemState)" you can say
"SysState, is this value in the group?". and it's static