back off of tailored creation
the idea I had working, with a stub record that cause a simplified 'refine' form was ok. but it had the disadvantage of possibly leaving behind an incomplete artwork record set if the user didn't proper cancel. So I would prefer to use logic to create the simpler forms during creation.
This could be done with new ARTWORK_CREATE states and more form selection logic, a system that is already in place.