stub in basic flow logic for renumbering
This gets almost all the ‘sequence of events’ logic in place to allow the user to get renumbering done.

They can request renumbering,
approve based on a summary of thier request,
request additional changes rather than approving,
and receive notification of successful or failed saves.

The only known bug is that on failed save, the last state of the request form is not right if there were additional requests made rather than an approval.