renumber stub code filled out with some logic
The action can now analize the request and create a set of entities for saving. It can create and display a summary of the moves requested and a description of errors it can’t resolve.

If there are no errors it will provides an Approval button. Otherwise, the user’s only option is to change the request and resubmit.