Segregate Helper Factory from Helper Parent class
Decompose classes to make a proper division of labor.

A factory was mingled with a parent class so the sub-classes the factory instantiated were also its own child classes… what??

All better now. The factory is a Lib class and the Helper parent class doesn’t have the factory code any more, just the default method implementations for the concrete helpers.