Update progress toward elimination of ArtStackElementHelper
Do a sweep of the elements to see how much work is left before the Helper can be deleted. LayerComponent took over all the legitimate tasks handled by this helper. Two elements still get piece table services from the helper, but this was never the proper place for that logic.

One proposed method is stubbed in to provide a form button to a couple of elements. This is both inappropriate for the helper and never resolved.

Once the pieces are handled elsewhere, this can be dumped.