Get EditionsController::assign template into normal stack-output path
This last view now follows the normal ‘layered’ output pattern through Artwork::review().

I stubbed in a way for EditionStack to return the Artwork object, something it didn’t do but which was required by the normal output pattern.

A better solution to this detail is required so, more changes to come. Possibly the EditionStack component will be compressed into a method on the ArtworkStack.