Add a new exception to the PieceFilter
At first I though returning an empty array in place of unknow pieces would plug this hole, and technically it would. But it then presents data based on an assumption rather that the true nature of the represented object. Pretty bad.

This Exception will guide development toward getting the proper data for the operation rather than just suppressing an error by tossing in a proxy value.