Make edition stack queries return new Providers object
Even thoug the stack query’s final home class is unknow, and the two current homes are overlap in function and are destined for deletion, making this change in them let’s me move the controller methods that use the query and the view code forward. This way, when the final home class and code is decided on, all the other parts will be set to go.