Button up the first-pass of custom finder creation for dispositions
The tool set is tentative, but it’s a workable start for queries on the date and type fields.

Next is likely to be an examination of how input and output sets of IDs will work. And along with this, some thought to the kind of data stacks that dispo record would control or participate in.