Create a first-draft class for managing id lists
The general idea is that we will need ID lists to do WHERE IN (list) queries. We could do this with containment, but in persuit of a flatter query style, we need to just do enough containment to get the IDs, then use them to query the next table.

This object accepts an Entity that has a property wich is a containment array which includes the IDs of interest. It can output the id list and other context and detail info