Get a sample page the works from bottom of stack to top
Example classes now exist that handle flat data sets and that allow them to be built from bottom (dispo) to top (artwork) but then let the output routines reassemble the data from top to bottom as a normal display.

This is a proof of concept. There are a lot of improvements that might be made.