enough tweaks to get the assignment for to render again
One hell of a hack job to do this. Provider was changed a while ago with an eye to compatablility with flat queries and composition from cached art stack data rather than being its own query/cache process.

So, mid-stream… a mess

The changes were surprisingly slight, mostly to extract the provider set (edition and formats) rather than expecting them to be exposed directly on the variable.

also, some stuff in the helper that seems related to changes in core entity property definition