Make assign process compatible with object not array
This is a quick and dirty fix for the changeover.

What I’ve done is make the object return the old, expected array in all cases. The real goal is to have the object eliminate some of the scattered loging in the different classes and files. That will take a more detailed study