Custom Field Type: Layer, first working draft
This is a working setup to automatically create Layer objects to wrap arrays of entities in the properties of a Stack type entity.

Everything is working automatically as it should. But setting up the associations of the tables, and the names of the properties to hold the Layer ojects was a bit clumsy.

I had to add the columns by hand in the _initializeSchema method. Not a surprise since the dummy table doesn’t have those fields. But when I also set the hasMany associations so that the tables were automatically available there was some kind of clash with the property names.

To solve this, I killed the Associations and wrote a magic getter for Table access.

I’m going to look for a cleaner way to get all this done. For now, this class does what its supposed to!