Change StackTrait into an Entity parent class
This is better as a parent class because the trait wasn’t going to be used for anything other than Entity sub classes. Technically, it could have been, but the in those cases the receiving class would have been fulfilling the role of an entity and should be written as one.