Open up LayerAccessInterface methods
This will now force the existance of the methods. The proposed ‘summarize’ interface methods have been moved in here too.

For now, implementation is a mess with a mix of Trait and Concrete code that is unexaminied, but mostly working.

There is a lot of refactoring to bring all this into harmony. For now, almost all tests are passing. The exception is StackEntity::distinct() and StackEntity::count()