Green StackEntity::distinct() tests
There is an error mode that can be triggered with a bad layer. For manual args I could make that just return an empty result. But for ArgObj i couldn’t.

But it isn’t the responsibility of code to stop bad program logic. So this would have to be a developer bug fix. At any rate, a bad layer value would yield bad data anyway.