Bring KeyValue() and var names into agreement
With the addition of distinct() and its delegation to valueList(), all the variable names and structure/style changed.

The third related method, keyValueList() got changes to bring it into harmony with the other methods.

They all now present string|object flexibility on the arguments and share large parts of their docblocks making them easier to understand.