Simplify ValueObject validity check
I was checking that the entity sent in to have its value extracted was of the type that the object was originally build to process. But the build process identified the type by the layer name while an actual entity from the layer was used in this check.

The layers don’t necessarily contain entities that share a name with that layer. In these cases the test failed.

Rather than create some elaborate new verification system, I removed that check (which seemed a bit over the top anyway)