fix the cause of an obscure bug
In one case, where an empty array was sent from a stack marshaller, the property ended up with an array rather than a layer.

Testing seemed to indicate this would not happen, but… this seems to fix whatever was going on.

Hash::check didn’t believe in an empty array node?