Fix the ‘default table’ problem
A new TableLocator has been defined in the controller to carry data to the Tables. But the default table generation for the controllers (eg: AritstsTable for ArtistsController) doesn’t use the Locator.

These defaults get built through an identified factory method and the default is on the default locator, not the currently configured one.

But the ModelAwareTrait allows configuration of the factory method. This code adds that and fixes the ‘defautl table’ problem.