add condition to change the way ManifestStack works
I’m implementing two kinds of Manifests now; ArtistManifests and ManagerManifests. There will be two classes and different data.

All current code assumes all Manifests to be Artist. This modified query makes sure those are the records returned for this class (which must be renamed).

Renaming this class and all tests, etc. will be done next in a different branch, but this change makes the current system act as it was without error.