Make PiecesController::renumber() work with new code
This is the final stub refactoring for renumber. The old tree-based EditionStackComponent::stackQuery() has been removed and ArtStack has been tooled to return the same values in the same arrangement as that old code.

The data-gathering lead-in code has been reworked to use the new tools.

But much of this is tentative. I just didn’t want to loose this pretty complicated action